Adding links to all Table of Contents headers
This adds links to all the Table of Contents to their corresponding
headers. Rather than using the generated markdown anchors, I added
explicit anchors with names, which clutters the markup a little bit,
but allows things to move around in the file without having to update
the links to match the new generated Markdown header anchor.

If the explicit headers are not preferable, I can switch this to the
generated ones. They look like: `#32-sealed-classes` and will change
based on the ordered list positioning.